requirements.txt
psutil
python-dotenv
tabulate
